FUSE-56 implement routing for setting page, remove drawer
