+++ apps/weight-loss-consultant-mobile/lib/constants/customer_bottom_navigator_index.dart
@@ -1,0 +1,7 @@
+enum CustomerBottomNavigatorIndex {
+  HOME,
+  CALENDAR,
+  SLIDE_UP,
+  MESSAGE,
+  SETTING
+}
