+++ apps/weight-loss-consultant-mobile/lib/pages/authentication/login_page.dart
@@ -162,7 +162,7 @@ class _LoginPageState extends State<LoginPage> {
                               );
                               dynamic result = await service.login();
                               if (result != null) {
-                                Navigator.pushReplacementNamed(context, RoutePath.customerHomePage);
+                                Navigator.pushNamedAndRemoveUntil(context, RoutePath.customerHomePage, (route) => false);
                               }
                             }
                           },
